requirements.txt
django==4.0.4
Pillow==9.1.0
django-email-users==1.1.0
signals==0.0.2